APIs/googleapis.com/games/v1/python-aiohttp-client/tests/test_scores_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.leaderboard_scores import LeaderboardScores
from openapi_server.models.player_leaderboard_score_list_response import PlayerLeaderboardScoreListResponse
from openapi_server.models.player_score_list_response import PlayerScoreListResponse
from openapi_server.models.player_score_response import PlayerScoreResponse
from openapi_server.models.player_score_submission_list import PlayerScoreSubmissionList


pytestmark = pytest.mark.asyncio

async def test_games_scores_get(client):
    """Test case for games_scores_get

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('includeRankType', 'include_rank_type_example'),
                    ('language', 'language_example'),
                    ('maxResults', 56),
                    ('pageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/games/v1/players/{player_id}/leaderboards/{leaderboard_id}/scores/{time_span}'.format(player_id='player_id_example', leaderboard_id='leaderboard_id_example', time_span='time_span_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_games_scores_list(client):
    """Test case for games_scores_list

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('timeSpan', 'time_span_example'),
                    ('language', 'language_example'),
                    ('maxResults', 56),
                    ('pageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/games/v1/leaderboards/{leaderboard_id}/scores/{collection}'.format(leaderboard_id='leaderboard_id_example', collection='collection_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_games_scores_list_window(client):
    """Test case for games_scores_list_window

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('timeSpan', 'time_span_example'),
                    ('language', 'language_example'),
                    ('maxResults', 56),
                    ('pageToken', 'page_token_example'),
                    ('resultsAbove', 56),
                    ('returnTopIfAbsent', True)]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/games/v1/leaderboards/{leaderboard_id}/window/{collection}'.format(leaderboard_id='leaderboard_id_example', collection='collection_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_games_scores_submit(client):
    """Test case for games_scores_submit

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('score', 'score_example'),
                    ('language', 'language_example'),
                    ('scoreTag', 'score_tag_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/games/v1/leaderboards/{leaderboard_id}/scores'.format(leaderboard_id='leaderboard_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_games_scores_submit_multiple(client):
    """Test case for games_scores_submit_multiple

    
    """
    body = {"scores":[{"score":"score","signature":"signature","kind":"kind","scoreTag":"scoreTag","leaderboardId":"leaderboardId"},{"score":"score","signature":"signature","kind":"kind","scoreTag":"scoreTag","leaderboardId":"leaderboardId"}],"kind":"kind"}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('language', 'language_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/games/v1/leaderboards/scores',
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

